backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
openai==0.28.1
pydantic==2.5.2
pydantic-settings==2.1.0
alembic==1.12.1
python-dateutil==2.8.2
nltk==3.8.1
scikit-learn==1.3.2
pandas==2.1.3
pytz>=2024.1 